feat: MAT-362 팀 매치 상세 패널 - 타입 정리
