+++ src/routes/teams/$teamId/matches/-components/MatchDetailCard/components/GoalList/index.ts
@@ -1,0 +1,1 @@
+export * from './GoalList';
