+++ src/routes/teams/$teamId/matches/-components/MatchDetailCard/components/MatchDetailLink/MatchDetailLink.tsx
@@ -4,9 +4,9 @@ import * as styles from './MatchDetailLink.css';
 
 export function MatchDetailLink() {
   return (
-    <div className={styles.container}>
+    <button className={styles.container}>
       <div className={styles.text}>자세히 보기</div>
       <ChevronRightIcon className={styles.icon} />
-    </div>
+    </button>
   );
 }
